+++ AWSAuthSDK/Sources/AWSMobileClient/AWSMobileClientExtensions.swift
@@ -568,7 +568,7 @@ extension AWSMobileClient {
                     return nil
                 }
                 if task.result != nil {
-                    // If global signout is successful, we attempt to revoke access tokens as signout locally.
+                    // If global signout is successful, we attempt to revoke access tokens and sign out locally.
                     self.revokeAccessTokensAndSignOutLocally(completionHandler: completionHandler)
                 } else if let error = task.error {
                     // If there is an error signing out globally, we notify the developer.
@@ -672,21 +672,27 @@ extension AWSMobileClient {
     ///
     /// - Parameters:
     ///   - completionHandler: completion handler for success or error callback.
-    public func deleteUser(signOut: Bool = true, completionHandler: @escaping ((Error?) -> Void)) {
+    public func deleteUser(completionHandler: @escaping ((Error?) -> Void)) {
         guard let currentActiveUser = self.userpoolOpsHelper.currentActiveUser else {
             let errorMessage = "Invalid CognitoUserPool configuration. This should not happen."
             completionHandler(AWSMobileClientError.invalidConfiguration(message: errorMessage))
             return
         }
-        let _ = currentActiveUser.delete().continueWith { (task) -> Any? in
+        let _ = currentActiveUser.delete().continueWith { [weak self] (task) -> Any? in
+            guard let self = self else {
+                let message = "Unexpectedly encountered nil when unwrapping self. This should not happen."
+                completionHandler(AWSMobileClientError.unknown(message: message))
+                return nil
+            }
             if task.result != nil {
                 // User was successfully deleted.
-                if signOut {
-                    // If signOut is true (default), perform global signout and invalidate tokens.
-                    let signOutOptions = SignOutOptions(signOutGlobally: true, invalidateTokens: true)
-                    self.internalSignOut(options: signOutOptions, completionHandler: completionHandler)
-                } else {
-                    completionHandler(nil)
+                // Attempt global signout, revoke access tokens, and sign out locally.
+                //
+                // In the case of hosted UI, launching a web session to remove the session cookie is
+                // unnessary because the user's account is no longer valid to receive new tokens.
+                let _ = currentActiveUser.globalSignOut().continueWith { _ in
+                    // Attempt to revoke access tokens and sign out locally.
+                    self.revokeAccessTokensAndSignOutLocally(completionHandler: completionHandler)
                 }
             } else if let error = task.error {
                 // If there is an error deleting the user, we notify the developer.
@@ -694,7 +700,6 @@ extension AWSMobileClient {
             }
             return nil
         }
-        return
     }
     
     internal func performUserPoolSuccessfulSignInTasks(session: AWSCognitoIdentityUserSession) {